moved bazaar config into inventory category.
