+++ src/main/java/at/hannibal2/skyhanni/config/ConfigUpdaterMigrator.kt
@@ -10,7 +10,7 @@ import com.google.gson.JsonPrimitive
 object ConfigUpdaterMigrator {
 
     val logger = LorenzLogger("ConfigMigration")
-    const val CONFIG_VERSION = 24
+    const val CONFIG_VERSION = 25
     fun JsonElement.at(chain: List<String>, init: Boolean): JsonElement? {
         if (chain.isEmpty()) return this
         if (this !is JsonObject) return null
